sliding doors added
baby steps, i added sliding doors.
sometimes i write notes here about things in my life for fun.
but not much is feeling very fun these days.
i do enjoy working on this project, and i hope that i am showing up for it more than i feel at times.
its important to have something to look forward to, something to focus on, something to do.
i should do the light triggers, or the music. but today the sliding doors felt like the thing to focus on.
lifes wild. i hope it changes as suddenly as it did for me this month but for the better.
and I hope adriana is having fun in los angeles.